requirements.txt
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
docutils==0.20.1
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
packaging==23.2
Pillow==10.0.1
pkginfo==1.9.6
Pygments==2.16.1
pyproject_hooks==1.0.0
pywin32-ctypes==0.2.2
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
twine==4.0.2
urllib3==2.0.6
zipp==3.17.0
